% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARIMAbuilder.R
\name{ARIMAbuilder}
\alias{ARIMAbuilder}
\title{ARIMA/SARIMA model builder}
\usage{
ARIMAbuilder(TS, date, p = 1, d = 1, q = 1, P = 1, D = 1, Q = 1, S = 12)
}
\arguments{
\item{TS}{The time series to build the models upon (remember to transform before using this function!)}

\item{date}{The date vector}

\item{p}{The maximum AR value, estimated from plot analysis}

\item{d}{The fixed differencing, often 0, 1 or 2}

\item{q}{The maximum MA value, estimated from plot analysis}

\item{P}{The maximum seasonal AR value, estimated from plot analysis}

\item{D}{The fixed seasonal differencing, often 0, 1 or 2}

\item{Q}{The maximum seasonal MA value, estimated from plot analysis}

\item{S}{The seasonality/period of the seasonal differencing}
}
\value{
list of top 5 ARIMA/SARIMA models
}
\description{
Lists the top 5 estimated ARIMA/SARIMA models of the form: ARIMA(p,d,q)x(P,D,Q), based on specified parameter values and ranges
}
