% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biasid.R
\docType{data}
\name{biasid}
\alias{biasid}
\title{Bias index for 2 different bias}
\format{
A numerical struct with dimensions 3x3x325
}
\usage{
biasid
}
\description{
A bias index containing bias indicies: 1 and 2, matching the synthetic bias data "TimeSeriesMatrixBias"
}
\keyword{datasets}
