% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biasidTMB.R
\docType{data}
\name{biasidTMB}
\alias{biasidTMB}
\title{Bias index for 1 bias}
\format{
A numerical struct with dimensions 3x3x325
}
\usage{
biasidTMB
}
\description{
A bias index containing bias index: 1, matching the synthetic bias data "TimeSeriesMatrixBiasTMB"
}
\keyword{datasets}
