% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEMOpllam.R
\docType{data}
\name{DEMOpllam}
\alias{DEMOpllam}
\title{Demonstration pl$lam}
\format{
A numerical vector with length: 325
}
\usage{
DEMOpllam
}
\description{
RW model estimates, calculated using the TMBTSA function on "TimeSeriesMatricBiasTMB" and "biasidTMB"
}
\keyword{datasets}
