% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEMOplsdlam.R
\docType{data}
\name{DEMOplsdlam}
\alias{DEMOplsdlam}
\title{Demonstration plsd$lam}
\format{
A numerical vector with length: 325
}
\usage{
DEMOplsdlam
}
\description{
RW model standard deviations, calculated using the TMBTSA function on "TimeSeriesMatricBiasTMB" and "biasidTMB"
}
\keyword{datasets}
