% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEMOtval.R
\docType{data}
\name{DEMOtval}
\alias{DEMOtval}
\title{DEMO T-value}
\format{
A numerical value
}
\usage{
DEMOtval
}
\description{
T-value for the 95% confidence interval for the TMBTSA function
}
\keyword{datasets}
