% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GridTSExtract.R
\name{GridTSExtract}
\alias{GridTSExtract}
\title{Grid Time Series Extractor}
\usage{
GridTSExtract(
  nco,
  lonid = "longitude",
  latid = "latitude",
  timeid = "date",
  measurementsid = "sea_level_anomaly",
  coord = c(-165, 74),
  radius = 0,
  dlon = 2,
  dlat = 4
)
}
\arguments{
\item{nco}{The NetCDF object (open with "nc_open" from the "ncdf4" library)}

\item{lonid}{The variable id for longitude vector. e.g.: "longitude"}

\item{latid}{The variable id for latitude vector. e.g.: "latitude"}

\item{timeid}{The variable id for the time vector. e.g.: "date"}

\item{measurementsid}{The variable id for the measurements vector. e.g.: "sea_level_anomaly"}

\item{coord}{The center coordinate for the TS grid area. e.g.: c(lon,lat) i.e. c(-164,74)}

\item{radius}{The square radius of the TS grid. e.g: 0=1x1 grid, 1=3x3 grid, 2=5x5 grid etc.}

\item{dlon}{Number of data points for each degree longitude}

\item{dlat}{Number of data points for each degree latitude}
}
\value{
Time series from grid
}
\description{
Extracts a time series from a NetCDF lon&lat coordinate grid
}
