% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{Time Series Analysis and Intermission Bias}
\usage{
hello(x)
}
\arguments{
\item{x}{The names for the hello function}
}
\value{
The output from \code{\link{print}}
}
\description{
Estimates a model based on time series analysis, and accounts for
intermission bias of satellite observations.
Input file type: NetCDF (Network Common Data Form).
}
\examples{
hello("Stine and Karina")
\dontrun{
hello("Mathias Høxbro")}
}
