% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IB.R
\name{IB}
\alias{IB}
\title{Intermission Bias}
\usage{
IB(TSM, date, biasvec = c(1, -2), biasid)
}
\arguments{
\item{TSM}{The time series struct with the intermission bias}

\item{date}{The time axis in the time series (dates, seconds, intervals)}

\item{biasvec}{The vector containging up to 2 different bias. If the data only contains 1 bias, insert that value instead of the vector}

\item{biasid}{A struct with the same dimensions as TSM, with integers indicating the bias: 0 for the restrained time series, 1 for the first bias and 2 for the second bias.}
}
\value{
Time series struct with bias correction
}
\description{
Corrects a struct of timeseries for up to two intermission biases
}
