% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeSeriesMatrixBias.R
\docType{data}
\name{TimeSeriesMatrixBias}
\alias{TimeSeriesMatrixBias}
\title{Time Series Matrix Bias}
\format{
A numerical struct with dimensions 3x3x325
}
\usage{
TimeSeriesMatrixBias
}
\description{
A time series matrix with added biases: 1 and -2, on the data: TSdata$TSmatrix
}
\keyword{datasets}
