% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeSeriesMatrixBiasTMB.R
\docType{data}
\name{TimeSeriesMatrixBiasTMB}
\alias{TimeSeriesMatrixBiasTMB}
\title{Time Series Matrix Bias for TMB}
\format{
A numerical struct with dimensions 3x3x325
}
\usage{
TimeSeriesMatrixBiasTMB
}
\description{
A time series matrix with added bias: 1, on the data: TSdata$TSmatrix
}
\keyword{datasets}
