% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS_area.R
\docType{data}
\name{TS_area}
\alias{TS_area}
\title{Time Series Area}
\format{
A numerical vector with length: 325
}
\usage{
TS_area
}
\description{
A time series containing the meaned and NaN removed time series, from using the TSnanrem function on TSdata$TSmatrix
}
\keyword{datasets}
