% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TSAIB.R
\name{TSAIB}
\alias{TSAIB}
\title{Time Series Analysis and Intermission Bias}
\usage{
TSAIB(x)
}
\arguments{
\item{x}{The names for the hello function}
}
\value{
The output from \code{\link{print}}
}
\description{
Estimates a model based on time series analysis, and accounts for
intermission bias of satellite observations.
Input file type: NetCDF (Network Common Data Form).
}
\examples{
TSAIB("Stine and Karina")
\dontrun{
TSAIB("Mathias")}
}
