% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TSAplots.R
\name{TSAplots}
\alias{TSAplots}
\title{Time Series Analysis Plots}
\usage{
TSAplots(x)
}
\arguments{
\item{x}{The time series for the TSAplots function}
}
\value{
The output from \code{\link{print}}
}
\description{
Estimates a model based on time series analysis, and accounts for
intermission bias of satellite observations.
Input file type: NetCDF (Network Common Data Form).
}
\examples{
TSAplots("Stine and Karina")
\dontrun{
TSAplots("Mathias")}
}
