% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TSdiagnostics.R
\name{TSdiagnostics}
\alias{TSdiagnostics}
\title{Time series diagnostics}
\usage{
TSdiagnostics(TS)
}
\arguments{
\item{TS}{The time series to be analyzed}
}
\value{
Basic statistics and characteristics
}
\description{
Shows basic statistics and characteristics of the time series data
}
