% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TSSmodel.R
\name{TSSmodel}
\alias{TSSmodel}
\title{Time Series model (simple constant + trend + sinus + cosine)}
\usage{
TSSmodel(TS, date, p = 12, t = 1, testsize = round(length(date)/10))
}
\arguments{
\item{TS}{The time series for the TSSmodel function}

\item{date}{The time axis in the time series (dates, seconds, intervals)}

\item{p}{The period. e.g.: p=12, for 12 months in a year}

\item{t}{The time steps. e.g.: t=1, for monthly time steps, with period 12}

\item{testsize}{The number of data points to be used for testing the model}
}
\value{
Simple TS model
}
\description{
Estimates a simple time series model
}
