% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TMBTSA.R
\name{TMBTSA}
\alias{TMBTSA}
\title{Template Model Builder Time Series Analysis}
\usage{
TMBTSA(TSM, date)
}
\arguments{
\item{TSM}{The time series matrix, with dimensions (x/lon,y/lat,timesteps)}

\item{date}{The time axis in the time series (i.e., dates, seconds, intervals)}
}
\value{
RW model
}
\description{
Utilizes TMB, which is a C++ interface for automatic differentiation that implements Laplace's method, to estimate a time series model
}
