% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TSAplots.R
\name{TSAplots}
\alias{TSAplots}
\title{Time Series Analysis Plots}
\usage{
TSAplots(TS, date, trans = "NONE", saveas = "NONE")
}
\arguments{
\item{TS}{The time series for the TSAplots function}

\item{date}{The time axis in the time series (dates, seconds, intervals)}

\item{trans}{data transform parameter: trans='log' for log transform, trans='sqrt' for sqrt transform, and trans='NONE' for no transform}

\item{saveas}{specifies the file format to save the plots in the current working directory. Possible formats is: PDF, JPEG, TIFF, BMP and PNG. e.g.: "pdf" or "jpeg". If none is specified, it will not save the plots}
}
\value{
Plots from TSAplots
}
\description{
Plots and saves the ACF and PACF for a pre-defined time series, with and without
data transformations (log, sqrt), and differencing.
Input type: Time series
}
