% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TSAplots.R
\name{TSAplots}
\alias{TSAplots}
\title{Time Series Analysis Plots}
\usage{
TSAplots(TS, date, trans)
}
\arguments{
\item{TS}{The time series for the TSAplots function}

\item{date}{The time axis in the time series (dates, seconds, intervals)}

\item{trans}{data transform parameter: trans='log' for log transform, trans='sqrt' for sqrt transform, and trans='NONE' for no transform}
}
\value{
Plots from TSAplots
}
\description{
Plots and saves the ACF and PACF for a pre-defined time series, with and without
data transformations (log, sqrt), and differencing.
Input type: Time series
}
